feat: md5 hashing, fix sliding window algorithm

+++ rx-unix/src/crc.h
@@ -1,8 +1,8 @@
 #ifndef CRC_H
 #define CRC_H
 
-unsigned long crc32(const void* buf, unsigned long size);
+unsigned int crc32(const void* buf, unsigned int size);
 
-unsigned char check_crc32(const void* buf, unsigned long size, const void* expected_crc_buf);
+unsigned char check_crc32(const void* buf, unsigned int size, const void* expected_crc_buf);
 
 #endif /* CRC_H */
